[EGD-7774] Fix compile on ARM GCC11
Fix compile on ARM GCC11

+++ module-bsp/board/rt1051/common/board.cpp
@@ -20,7 +20,7 @@ extern "C"
 
 #include <cstdint>
 
-extern std::uint8_t __sdram_cached_start[];
+extern std::uint32_t __sdram_cached_start[];
 
 namespace bsp
 {
